backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
PyPDF2
python-multipart
requests
openai
youtube-transcript-api
langchain
psycopg2
sqlalchemy
